test: OpenMarketService Get API 테스트 코드 구현 #3
